# types/json: replace binary.Read with binary.BigEndian.Uint16